Replace Promise handling in actions with middleware
This adds a fetchJSONMiddleware method that is applied on the store. If
an action is dispatched with a `promise` property it will be handled
properly given the JSON API. This allows the actions to be cleaned up
considerably as the repetitive logic for handling Promise resolution is
no longer duplicated in each action.