Improve middleware and add new unit form
Fixed up the fetchJSONMiddleware's error handling. Responses with a
status code representing an error tried to access the error message
directly from a Promise instead of an actual JSON object. Tired mistake.
Fixed up now.

Stateless UnitForm component with a NewUnitContainer to handle the
current attributes of the new unit, in addition to the current time
delta, etc.