Fix mark complete bug
I think the application is experiencing some microsecond differences
between client fetch and server, such that when attempting to mark a
unit complete, the server spits back an error because it is not quite
ready to be completed.

Instead of checking if the delta is less than 0, check if it's 100
microseconds past complete.